backend/requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gitdb==4.0.12
GitPython==3.1.41
mysql==0.0.3
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.9.0
setuptools==75.6.0
smmap==5.0.2
sqlparse==0.5.3
timedelta==2020.12.3
